Refactor timeline view and component AB#14745 (#4739)
- moves dataset handling out of TimelineView (and into the component)
- refactors timeline component to support HDID and datasets passed as props
  - passes HDID from the view to the component and then to the timeline entry components for each dataset
  - removes noisy logging
  - consolidates and eliminates some getters and functions
- standardizes names of entry types and comment entry types
- removes custom names for timeline entry components
- renames TimelineView to UserTimelineView
- renames LinearTimelineComponent to TimelineComponent
